Working table insertion

+++ backend/go-code/internal/random/random.go
@@ -1,6 +1,6 @@
 package random
 
-// Cryptographically Random Unsigned Integers
+// Cryptographically Random Integers
 
 import (
     "crypto/rand"
@@ -12,12 +12,12 @@ const (
     limit32 int64 = max32 + 1
 )
 
-func Random32() uint32 {
+func Random32() int32 {
     exclusiveMax := big.NewInt(limit32)
     id, _ := rand.Int(rand.Reader, exclusiveMax)
-    return uint32(uint64(id.Int64()))
+    return int32(id.Int64())
 }
 
-func Random64() uint64 {
-    return uint64(Random32()) << 32 + uint64(Random32())
+func Random64() int64 {
+    return int64(Random32()) << 32 + int64(Random32())
 }
